# docs/api-specification.yaml
openapi: 3.1.0
info:
  title: AI Part Designer API
  description: |
    RESTful API for AI-powered 3D part design and generation platform.
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the 
    `Authorization` header as `Bearer <token>`.
    
    ## Rate Limiting
    - Free tier: 30 requests/minute
    - Pro tier: 100 requests/minute
    - Enterprise tier: 500 requests/minute
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
    
    ## Versioning
    API version is specified in the URL path: `/api/v1/`
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@aipartdesigner.com
  license:
    name: Proprietary
    url: https://aipartdesigner.com/terms
servers:
  - url: https://api.aipartdesigner.com/api/v1
    description: Production server
  - url: https://api.staging.aipartdesigner.com/api/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Local development
...
paths:
...
  /designs/{designId}/versions:
    get:
      tags: [Designs]
      summary: List design versions
      operationId: listDesignVersions
      security:
        - bearerAuth: []
      parameters:
        - name: designId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Version history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DesignVersion'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DesignVersion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        versionNumber:
          type: integer
        fileUrl:
          type: string
          format: uri
        thumbnailUrl:
          type: string
          format: uri
        fileFormats:
          type: object
          additionalProperties:
            type: string
            format: uri
        parameters:
          type: object
        geometryInfo:
          type: object
          properties:
            boundingBox:
              type: object
              properties:
                x:
                  type: number
                y:
                  type: number
                z:
                  type: number
            volume:
              type: number
            surfaceArea:
              type: number
            triangleCount:
              type: integer
            isManifold:
              type: boolean
        changeDescription:
          type: string
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []